Percival Carrier Simulator now a class
Can run either synchronously or asynchronously forever or just once. Unittest friendly class.
